requirements.txt
# Core AI
google-generativeai>=0.8.0    # Gemini API
google-adk>=0.1.0             # Google Agent Development Kit

# Utilities
python-dotenv>=1.0.0          # Environment variable management
pytest>=8.0.0
